Python/requirements.txt
opencv
pillow
matplotlib